renamed to match intended purpose + misc formatting
